+++ weave-js/src/common/types/list.ts
@@ -1,0 +1,6 @@
+import {Type} from '@wandb/weave/core';
+
+export const LIST_ANY_TYPE: Type = {
+  type: 'list' as const,
+  objectType: 'any' as const,
+};
